wip dynamodb resource - refactor common functions
