+++ Cargo.toml
@@ -96,3 +96,5 @@ ttl_cache = "0.5.1"
 utoipa = { version = "3.2.1", features = [ "uuid", "chrono" ] }
 utoipa-swagger-ui = { version = "3.1.3", features = ["axum"] }
 uuid = "1.2.2"
+aws-sdk-dynamodb = "0.27.0"
+aws-config = "0.55.2"
