+++ provisioner/Cargo.toml
@@ -7,9 +7,9 @@ description = "Service responsible for provisioning and managing resources for s
 publish = false
 
 [dependencies]
-aws-config = "0.55.2"
+aws-config = { workspace = true }
 aws-sdk-rds = "0.27.0"
-aws-sdk-dynamodb = "0.27.0"
+aws-sdk-dynamodb = { workspace = true }
 aws-sdk-iam = "0.27.0"
 aws-sdk-sts = "0.27.0"
 clap = { workspace = true, features = ["env"] }
